Move moderation log stream to global settings
When using "skip_existing = 1" when generating a stream, the first results returned from the API are disregarded. Since we re-created the stream every time we check for updates, we almost never see updates.

Moving this to global settings allows us to keep calling the same generator and get just the new results.